refactor: remove ZRect. This was insufficient.
Instead allow adding a FocusFlag multiple times and replace the current z_areas() with a simple area_z() -> u16.

Add the area_z() for container flags.

Enable stacking the z-value of containers. All widgets added to a container will get at least the z of the container. A sub container adds/stacks it's z-value on top of the surrounding container.

refactor: Focus:

Add the current z_base to the builder and implement stacking.
Remove the current z_areas and replace them with a new Vec<bool> to indicate duplicates. The duplicate flag is only used in __start_change() to avoid resetting the flags twice and rendering them useless this way.

In all other cases the first occurrance wins, which is fine. Additionaly duplicates are only allowed to be Navigation::Mouse which prevents a lot of cases.