major: Replaced the current Message type variable with Event. This offloads the complete event-type to the application.
* no more crossterm dependence.
* extra Pollxx can be added and use the standard event handling. An extra Pollxxx can require support for its own events.
* as a side Timers and ThreadPool can be deactivated completely. They need a few bytes still, but that's ok. Context functions just panic if you call them without underlying support. Which is ok.
* adding a new Pollxx is very straightforward now.

The change was surprisingly easy though. Most of the changes come from the optional timer and threading.

The AppState trait is nice and clean now. It retains init() and gains the lost shutdown(). Which is probably for the best to keep those out of regular event-handling. And I keep the error() functions, but this one is always defaulted except for the primary AppState anyway.

*** Upgrading ***

Is copypasta more or less.

One more advantage here: event forwarding happens only in the event() function and is not distributed over multiple functions any more. That has been the source of many quirks.

*** App code-sizes ***

Stay the same. Maybe minor reductions. Ultra goes slightly over a 100 lines, but I could bring it back if I leave out threading and timers. ... Done 99!!