fix: #5
Add a flag for scroll_to_cursor. This calculates the offset needed during rendering to ensure a visible cursor with the actual bounds. This is very necessary during initialization to set a first cursor position.

This enables scrolling the cursor into view even before the first render. Should probably help during resize but I have not seen any problem with that.

Any function that changes the offset resets this flag.

fix: #4
Hide the cursor if there is a selection.

maybe break:
Remove the limitations on ScrollState::set_offset() for what is a valid offset. As established before, any combination of scroll-parameters is valid and must be supported by widgets.